improve the contract deployment error message
